Move `SenExprBuilder` to a header.
No functional change intended.
